Burningsvideo 0.52 Linux Warnings
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@6155 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CSoftwareTexture2.h
@@ -55,7 +55,7 @@ public:
 	u32 getMipmapLevel(s32 newLevel) const
 	{
 		if ( newLevel < 0 ) newLevel = 0;
-		else if ( newLevel >= array_size(MipMap)) newLevel = array_size(MipMap) - 1;
+		else if ( newLevel >= (s32)array_size(MipMap)) newLevel = array_size(MipMap) - 1;
 
 		while ( newLevel > 0 && MipMap[newLevel] == 0 ) newLevel -= 1;
 		return newLevel;
